Sonar fixes

+++ src/creatures/creature.cpp
@@ -644,7 +644,7 @@ void Creature::onDeath()
 
 	const int64_t timeNow = OTSYS_TIME();
 	const uint32_t inFightTicks = g_configManager().getNumber(PZ_LOCKED);
-	int32_t mostDamage = 0;
+	int64_t mostDamage = 0;
 	std::map<Creature*, uint64_t> experienceMap;
 	for (const auto& it : damageMap) {
 		if (Creature* attacker = g_game().getCreatureByID(it.first)) {
